Delete not needed css and fix grunt path
The CSS was not needed.  The !important rule was added because it was not taking effect and the reason why was the selector was in another rule right below.

The image path in grunt for datatables images was wrong when running grunt.  The images for datatables should point to the images folder in the css folder since datatables has those paths in the css file.